add nominal.core.filetype
