+++ nominal/core/__init__.py
@@ -5,6 +5,7 @@ from nominal.core.checklist import Check, Checklist, ChecklistBuilder
 from nominal.core.client import NominalClient
 from nominal.core.connection import Connection
 from nominal.core.dataset import Dataset, poll_until_ingestion_completed
+from nominal.core.filetype import FileType, FileTypes
 from nominal.core.log import Log, LogSet
 from nominal.core.run import Run
 from nominal.core.user import User
@@ -20,6 +21,8 @@ __all__ = [
     "ChecklistBuilder",
     "Connection",
     "Dataset",
+    "FileType",
+    "FileTypes",
     "Log",
     "LogSet",
     "NominalClient",
